EE-27518 Show errors in Welsh.
You may have notied that this is a somewhat half-baked solution.  The reason for this
is that to make the proper fix, we would have to add language support to the nginx
proxy maintained by ACP which looks like a much bigger job.  As the chances of the
error page is tiny (as judged by usage statistics for the previous enquiry form)
it has been decided to go with this simple solution instead.